requirements.txt
torch==2.8.0
torchvision
open-clip-torch>=2.24.0
faiss-gpu-cu12
pandas>=1.5.0
numpy>=1.24.0
pillow>=9.5.0
tqdm>=4.65.0
scipy>=1.10.0
scikit-learn>=1.2.0
transformers==4.57.3
peft>=0.15.2
sentencepiece>=0.1.99
accelerate>=0.28.0
einops>=0.7.0
timm>=0.9.12
qwen-vl-utils==0.0.14
scikit-image
opencv-contrib-python
pyproj
img2dataset==1.47.0
